refactor: :recycle: DRY: use rollup to create commonjs modules from esm
