+++ lib/fingerprint.browser.mjs
@@ -1,4 +1,4 @@
-var pad = require('./pad.js');
+import pad from './pad';
 
 var env = typeof window === 'object' ? window : self;
 var globalCount = 0;
@@ -10,6 +10,6 @@ var clientId = pad((mimeTypesLength +
   navigator.userAgent.length).toString(36) +
   globalCount.toString(36), 4);
 
-module.exports = function fingerprint () {
+export default function fingerprint () {
   return clientId;
-};
+}
