+++ lib/fingerprint.mjs
@@ -1,4 +1,4 @@
-import pad from './pad.mjs';
+import pad from './pad';
 import os from 'os';
 
 function getHostname () {
