+++ lib/fingerprint.react-native.mjs
@@ -1,8 +1,8 @@
-var pad = require('./pad.js');
+import pad from './pad';
 
 var globalCount = Object.keys(global);
 var clientId = pad(globalCount.toString(36), 4);
 
-module.exports = function fingerprint () {
+export default function fingerprint () {
   return clientId;
-};
+}
