--- lib/pad.js
@@ -1,4 +1,0 @@
-module.exports = function pad (num, size) {
-  var s = '000000000' + num;
-  return s.substr(s.length - size);
-};
