+++ package.json
@@ -6,30 +6,31 @@
     "name": "Bugsnag",
     "url": "https://bugsnag.com"
   },
-  "main": "index.js",
+  "main": "dist/index.js",
   "types": "cuid.d.ts",
   "exports": {
     ".": {
-      "import": "./index.mjs",
-      "default": "./index.js"
+      "import": "./index.esm.mjs",
+      "require": "./dist/index.js"
     },
     "./lib/fingerprint": {
-      "browser": "./lib/esm/fingerprint.browser.mjs",
-      "node": "./lib/esm/fingerprint.mjs",
-      "default": "./lib/fingerprint.js"
+      "browser": "./lib/fingerprint.browser.mjs",
+      "node": "./lib/fingerprint.mjs",
+      "default": "./dist/lib/fingerprint.js"
     }
   },
   "imports": {
     "#fingerprint": {
-      "browser": "./lib/esm/fingerprint.browser.mjs",
-      "default": "./lib/esm/fingerprint.mjs"
+      "browser": "./lib/fingerprint.browser.mjs",
+      "node": "./lib/fingerprint.mjs",
+      "default": "./lib/fingerprint.react-native.js"
     }
   },
   "browser": {
-    "./lib/fingerprint.js": "./lib/fingerprint.browser.js"
+    "./dist/lib/fingerprint.js": "./dist/lib/fingerprint.browser.js"
   },
   "react-native": {
-    "./lib/fingerprint.js": "./lib/fingerprint.react-native.js"
+    "./dist/lib/fingerprint.js": "./dist/lib/fingerprint.react-native.js"
   },
   "keywords": [
     "id",
@@ -44,12 +45,13 @@
   },
   "files": [
     "lib",
+    "dist",
     "cuid.d.ts",
-    "index.js",
-    "index.mjs"
+    "index.esm.mjs"
   ],
   "license": "MIT",
   "devDependencies": {
+    "@rollup/plugin-alias": "^5.1.1",
     "browserify": "17.0.0",
     "eslint": "^8.49",
     "jasmine": "^2.5.0",
@@ -59,13 +61,15 @@
     "karma-chrome-launcher": "^2.2.0",
     "karma-jasmine": "^1.1.1",
     "mkdirp": "^0.5.1",
+    "rollup": "^4.34.2",
     "uglify-js": "^2.7.4",
     "watchify": "^4.0.0",
     "webworkify": "^1.4.0"
   },
   "scripts": {
-    "build": "mkdirp dist && browserify index.js -s cuid -o dist/cuid.js && uglifyjs dist/cuid.js -c -m -o dist/cuid.min.js",
-    "lint": "eslint index.js lib test",
+    "prebuild": "rm -rf dist",
+    "build": "rollup -c",
+    "lint": "eslint index.mjs lib test",
     "test": "npm run lint && npm run test:server && npm run test:browser",
     "test:browser": "karma start test/karma.conf.js",
     "test:server": "jasmine test/**/*.js",
